[#25]:svarga:iex, consolidated all iex code into a single header
fix iex
